REFACT: todo page에서 리스트 관리 로직 훅으로 분리
